Add additional safeguards for logic related to the change of the pacman repo name. Fail when pacman.conf.tmpl doesn't exist, init the new repo before building a package, regenerate the makepkg Docker image if a change of the repo name happened.

+++ startup.sh
@@ -1,13 +1,16 @@
 #!/usr/bin/env bash
 
 if [ ! -f data/settings.ini ]; then
-    cp settings.ini.template data/settings.ini
-    echo "settings.ini did not exist, copied from settings.ini.template"
+    echo "settings.ini did not exist, copyingfrom settings.ini.template"
+    cp settings.ini.template data/settings.ini || { echo "settings.ini.template missing."; exit 1; }
 fi
 
 source <(grep pacmanrepo_name data/settings.ini | tr -d ' ')
 export pacmanrepo_name
+
+[ -f pacman.conf.tmpl ] || { echo 'pacman.conf.tmpl is missing.'; exit 1; }
 envsubst '$pacmanrepo_name' < pacman.conf.tmpl > /etc/pacman.conf
+[ -f makepkg/docker/pacman.conf.tmpl ] || { echo 'makepkg/docker/pacman.conf.tmpl is missing.'; exit 1; }
 envsubst '$pacmanrepo_name' < makepkg/docker/pacman.conf.tmpl > makepkg/docker/pacman.conf
 
 python manage.py migrate
